Reload leaderboards when props updated
Problem

In your app, if you change props based on a toggle or some other
change (e.g. on toggle, show teams, otherwise, show individuals), the
leaderboard won't update. This is because it only fetches data on
componentDidMount

Solution

Add a componentDidUpdate hook, which will refresh the leaderboard if the
props you pass it have changed