bug dealing with UNIX timestamps
JG returns UNIX time as string wrapped as '\Date(`. Moment was able to deal with this syntax automatically, but for day.js we have to strip out the extra formatting and just include the actual UNIX timestamp as a digit, otherwise we get a Type Error. Also, it does return true as a day object, but just evaluates as string 'Invalid Date' (this is what gives type error when we try to evaluate and format this). So, we first have to test if the day obj is valid, if then not format the string to proper UNIX digit before converting to ISO string.